% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanglogitMer-compare.R
\name{bridge_sampler.stanglogitFit}
\alias{bridge_sampler.stanglogitFit}
\title{bridge_sampler for stanglogitFit objects}
\usage{
bridge_sampler.stanglogitFit(object, ...)
}
\arguments{
\item{object}{a stanglogitFit object}
}
\value{
an ""psis_loo" "loo" object
}
\description{
bridge_sampler for stanglogitFit objects
}
