% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanglogitMer-utilities.R
\name{.findMaxMin}
\alias{.findMaxMin}
\title{Internal function to find maximum and minimum of the dependent variable divided by cell.}
\usage{
.findMaxMin(formula, data, dependent)
}
\arguments{
\item{formula}{the formula to split the dependent variable into cells.}

\item{data}{the original data frame.}
}
\description{
\code{findMaxMin} returns a data frame with the same number of rows
of the original data frame with two columns: the minimum and
the maximum value obtained by the dependent variable
grouped per cell.
}
