% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanglogitMer-utilities.R
\name{.paste.formulas}
\alias{.paste.formulas}
\title{Internal function to write formulas from a vector of individual strings.}
\usage{
.paste.formulas(x)
}
\arguments{
\item{x}{is the vector of individuals strings. The first string is the dependent variable.}
}
\description{
Internal function to write formulas from a vector of individual strings.
}
\examples{
paste.formulas(c("y","Condition","Group"))
}
