% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanglogitMer-functions.R
\name{glogit}
\alias{glogit}
\title{Generalized Logit Function.}
\usage{
glogit(x, a, k, g, s)
}
\arguments{
\item{x}{the independent continuous variable.}

\item{a}{the minimum value.}

\item{k}{the maximum value.}

\item{g}{the growth parameter.}

\item{s}{the shift (intercept) parameter.}
}
\value{
The value from the generalized logit function.
}
\description{
\code{glogit} returns the value on the generalized logit curve,
along the covariate \code{x}, between the \code{a}
and \code{k} parameters, with growth \code{g} and shift \code{s}.
}
\details{
\deqn{a + \frac{k-a}{1+e^{-g \times (x-s)}}}
}
\examples{
glogit(-100:100,-70,100,0.5,10)
glogit(-100:100,0,1,0.5,10)
plot(x=-100:100,y=glogit(-100:100,-70,100,0.2,-10),type="l")
}
\seealso{
\code{\link{invglogit}}
}
