% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanglogitMer-functions.R
\name{invglogit}
\alias{invglogit}
\title{Inverse Generalized Logit Function.}
\usage{
invglogit(y, a, k, g, s)
}
\arguments{
\item{y}{the dependent variable / a variable coming from a Generalized Logit Function.}

\item{a}{the minimum value.}

\item{k}{the maximum value.}

\item{g}{the growth parameter.}

\item{s}{the shift (intercept) parameter.}
}
\value{
The value from the inverse generalized logit function.
}
\description{
\code{invglogit} returns the value from the inverse generalized logit function,
along the continuous \code{y}, coming from a Generalized Logit Function, between the \code{a}
and \code{k} parameters, with growth \code{g} and shift \code{s}.
}
\details{
\deqn{a + \frac{k-a}{1+e^{-g \times (x-s)}}}
}
\examples{
y = glogit(-100:100,-70,100,0.5,10)
x = invglogit(y,-70,100,0.5,10)
plot(x=x,y=y,type="l")
}
\seealso{
\code{\link{glogit}}
}
