% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanglogitMer-compare.R
\name{loo.stanglogitFit}
\alias{loo.stanglogitFit}
\title{loo for stanglogitFit objects}
\usage{
loo.stanglogitFit(object, ...)
}
\arguments{
\item{object}{a stanglogitFit object}
}
\value{
an ""psis_loo" "loo" object
}
\description{
loo for stanglogitFit objects
}
