% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanglogitMer-plot.R
\name{plot.stanglogitFit}
\alias{plot.stanglogitFit}
\title{plot for stanglogitFit objects}
\usage{
plot.stanglogitFit(object, pars = NULL, type = "combo", ...)
}
\arguments{
\item{object}{a stanglogitFit object}

\item{pars}{an optional vector of strings, with possible values:
\describe{
    \item{all} all the parameters
    \item{growth} only the growth parameters
    \item{shift} only the shift parameters
    \item{random} only the random variances
    \item{cor_random} only the correlations between random variables
    \item{sigma} only the overall sigma of the model
}}

\item{type}{to select the type of plot:
\item{combo} a density and a trace plot
\item{area} a density plot
\item{hist} a density histogram
\item{intervals} an 95% intervals plot
\item{pairs} a pairs plot (only with \code{bayesplot})
\item{trace} a trace/catterpillar plot
\item{violin} a violin plot (only with \code{bayesplot})}

\item{...}{other bayesplot or ggplot2 arguments}
}
\value{
a plot, a ggplot2 object, or a bayesplot object
}
\description{
plot for stanglogitFit objects
}
