% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanglogitMer-plot.R
\name{plot_gaussian}
\alias{plot_gaussian}
\title{plot_gaussian for stanglogitFit objects}
\usage{
plot_gaussian(object, CI = 0.95)
}
\arguments{
\item{object}{a stanglogitFit object}

\item{CI}{the dimension of the credible interval taken into account. Default 0.95.}
}
\value{
a ggplot2 object or a base-R plot
}
\description{
\code{plot_gaussian()} plots the resulting generalised logit curves from the posterior distributions of the model,
as gaussian curves, with mean the Shift parameter, and standard deviation the inverse of the
Growth parameter.
For each combination of population-level factors for the Growth and Shift parameters, 9 curves will be plotted.
One curve will come from the median value of the posterior distributions of the Growth and Shift parameters,
the other eight parameters will be obtained by the combination of the boundaries of the Credible
Interval for the Growth and Shift parameters. The dimension of the Credible Interval can be
set by means of the \code{CI} parameter.
}
\seealso{
\code{\link{plot_sigmoid}}
}
