% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanglogitMer-plot.R
\name{plot_sigmoid}
\alias{plot_sigmoid}
\title{plot_sigmoid for stanglogitFit objects}
\usage{
plot_sigmoid(object, CI = 0.95)
}
\arguments{
\item{object}{a stanglogitFit object}

\item{CI}{the dimension of the credible interval taken into account. Default 0.95.}
}
\value{
a ggplot2 object or a base-R plot
}
\description{
\code{plot_sigmoid()} plots the resulting generalised logit curves from the posterior distributions of the model.
For each combination of population-level factors for the Growth and Shift parameters, 9 curves will be plotted.
One curve will come from the median value of the posterior distributions of the Growth and Shift parameters,
the other eight parameters will be obtained by the combination of the boundaries of the Credible
Interval for the Growth and Shift parameters. The dimension of the Credible Interval can be
set by means of the \code{CI} parameter.
}
\details{
A sigmoid function can be seen as a cumulative gaussian function. In order to plot the gaussian curves,
use \code{plot_gaussian()}.
}
\seealso{
\code{\link{plot_gaussian}}
}
