% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanglogitMer-functions.R
\name{summary.stanglogitFit}
\alias{summary.stanglogitFit}
\title{Summary for stanglogitFit objects}
\usage{
\method{summary}{stanglogitFit}(object)
}
\arguments{
\item{object}{a stanglogitFit object}
}
\value{
an "summary.stanglogitFit" object
}
\description{
Summary for stanglogitFit objects
}
