% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanglogitMer-compare.R
\name{waic.stanglogitFit}
\alias{waic.stanglogitFit}
\title{waic for stanglogitFit objects}
\usage{
waic.stanglogitFit(object, ...)
}
\arguments{
\item{object}{a stanglogitFit object}
}
\value{
an ""psis_loo" "waic" object
}
\description{
waic for stanglogitFit objects
}
